Refactor Additional Task 2
